+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog].
 - Added localization for common config key `pigeon.allow_self_delivery`. ([#8])
 - Naming Pigeon with the name of a very specific fictional character will alter its appearance,
   inspired by said very specific fictional character. ([#9])
+- Temporarily use sounds from vanilla parrots for Pigeons. ([#3])
 
 ### Changed
 
@@ -103,6 +104,7 @@ The format is based on [Keep a Changelog].
 [#17]: https://github.com/GiienaAstrella/pigeon-chat/issues/17
 [#18]: https://github.com/GiienaAstrella/pigeon-chat/issues/18
 [#2]: https://github.com/GiienaAstrella/pigeon-chat/issues/2
+[#3]: https://github.com/GiienaAstrella/pigeon-chat/issues/3
 [#4]: https://github.com/GiienaAstrella/pigeon-chat/issues/4
 [#5]: https://github.com/GiienaAstrella/pigeon-chat/issues/5
 [#7]: https://github.com/GiienaAstrella/pigeon-chat/issues/7